Controller refactor, recipe_handler fixes
Adds a Controller base class that can be inherited in controller classes to
dynamically set up the environment such as loading default ROS Params,
registering callbacks to set points and measurements ROS Topics, and publishing
commanded output values.

recipe_handler:
Fixes the bug that caused the node to crash when the recipe was
stopped by a ROS service call (via the UI or via `rosservice call`).
Moved the methods to set/clear the ROS Parameters in the setter/getter as
stopping could be called via the ROS Service.
Also fix the desired topic to conform to the new format `/<variable>/desired`.

I assume the default topics the Controllers were publishing and subscribing to
were part of a previous implementation as I couldn't find any code that
referenced them, so I removed them and now the module raises an exception if no
`"variable"` is set.