Fix recipe handler when ROS Service stop_recipe is called
1. Moved the `current_recipe`-related ROS Param setters to the get/set with lock
to ensure that stopping a recipe asynchronously via the `stop_recipe` ROS
service also clears them as well (compared to just clearing them when the recipe
organically ends)

2. Fixes a crash that occurs when calling the `stop_recipe` ROS service if a
recipe is currently running.

It is happening because the  `for` that iterates over the recipe setpoints has
to `break`, then `self.clear_recipe()` is called, but the private `self._recipe`
has already been cleared by the asynchronous ROS service call, so a
`RecipeIdleError("No recipe is running")` is raised.